[location] move to global/exception
