+++ src/main/java/HDmedi/Server/global/exception/GlobalExceptionHandler.java
@@ -1,7 +1,5 @@
 package HDmedi.Server.global.exception;
 
-import HDmedi.Server.domain.user.dto.response.ErrorResponse;
-
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
